+++ krampoline/public/data/todos.json
@@ -1,17 +1,37 @@
 [
   {
     "id": "0",
-    "title": "밥먹기",
-    "status": "active"
+    "content": "밥먹기",
+    "status": "active",
+    "startDate": "2023-10-02",
+    "endDate": "2023-10-02"
   },
   {
     "id": "1",
-    "title": "숨쉬기",
-    "status": "active"
+    "content": "숨쉬기",
+    "status": "active",
+    "startDate": "2023-09-03",
+    "endDate": "2023-10-03"
   },
   {
     "id": "2",
-    "title": "놀기",
-    "status": "active"
+    "content": "놀기",
+    "status": "active",
+    "startDate": "2023-04-05",
+    "endDate": "2023-05-15"
+  },
+  {
+    "id": "3",
+    "content": "공부하기",
+    "status": "active",
+    "startDate": "2023-10-27",
+    "endDate": "2023-11-29"
+  },
+  {
+    "id": "4",
+    "content": "청소하기",
+    "status": "active",
+    "startDate": "2023-10-27",
+    "endDate": "2023-10-27"
   }
 ]
