+++ krampoline/src/components/Task.tsx
@@ -3,7 +3,7 @@ import Checkbox from "@mui/material/Checkbox";
 import { PiTrashLight, PiPencilLight } from "react-icons/pi";
 
 function Task({ todo, onUpdate, onDelete }) {
-  const { id, title, status } = todo;
+  const { id, content, status } = todo;
 
   const handleChange = (e) => {
     onUpdate({ ...todo, status: e.target.checked ? "completed" : "active" });
@@ -19,12 +19,17 @@ function Task({ todo, onUpdate, onDelete }) {
           id={id}
           checked={status === "completed"}
           onChange={handleChange}
-          defaultChecked
           color="secondary"
           className="cursor-pointer"
         />
-        <label htmlFor={id} className="flex justify-left items-center">
-          {title}
+        <label
+          htmlFor={id}
+          className="flex justify-left items-center"
+          style={{
+            textDecoration: status === "completed" ? "line-through" : "none",
+          }}
+        >
+          {content}
         </label>
       </div>
       <span className="flex justify-right items-center">
